# app/src/main/java/org/skillhub/model/Image.java
package org.skillhub.model;


import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class Image {

  private String name;
  private String type;
  private byte[] data;

}