# app/src/main/java/org/skillhub/model/Skill.java
package org.skillhub.model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.io.Serializable;
import java.math.BigDecimal;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class Skill implements Serializable {
  private String name;
  private BigDecimal score;
}